+++ packages/node/CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `--network-endpoint-config` add optional parameter `pageLimit`
+
+### Fixed
+- Performance issue caused by making too many RPC requests.
+
 ## [4.4.0] - 2025-02-05
 ### Changed
 - Update stellar sdk to v13 (#101)